refactor(api-docs): update imports
