+++ plugins/api-docs/src/components/ApiEntityPage/ApiEntityPage.tsx
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import { ApiEntityV1alpha1, Entity } from '@backstage/catalog-model';
+import { ApiEntity, Entity } from '@backstage/catalog-model';
 import {
   Content,
   errorApiRef,
@@ -25,14 +25,13 @@ import {
   Progress,
   useApi,
 } from '@backstage/core';
-// TODO: Circular ref
 import { catalogApiRef } from '@backstage/plugin-catalog';
 import { Box } from '@material-ui/core';
 import { Alert } from '@material-ui/lab';
 import React, { useEffect } from 'react';
 import { useNavigate, useParams } from 'react-router-dom';
 import { useAsync } from 'react-use';
-import { ApiDefinitionCard } from '../ApiDefinitionCard/ApiDefinitionCard';
+import { ApiDefinitionCard } from '../ApiDefinitionCard';
 
 const REDIRECT_DELAY = 1000;
 function headerProps(
@@ -125,7 +124,7 @@ export const ApiEntityPage = () => {
       {entity && (
         <>
           <Content>
-            <ApiDefinitionCard apiEntity={entity as ApiEntityV1alpha1} />
+            <ApiDefinitionCard apiEntity={entity as ApiEntity} />
           </Content>
         </>
       )}
